requirements.txt
# Core Django
Django==4.2.10
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2024.2
pytz==2025.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
django-allauth==0.61.1
django-axes==6.3.0
django-guardian==2.4.0
django-otp==1.5.4
PyJWT==2.9.0
cryptography==43.0.3

# API & REST
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.1.0rc8
django-cors-headers==4.3.1
django-filter==23.5
coreapi==2.3.3
coreschema==0.0.4

# Frontend & UI
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-widget-tweaks==1.5.0
django-htmx==1.21.0
Jinja2==3.1.6
MarkupSafe==3.0.2

# Storage & Media
django-storages==1.14.2
django-cloudinary-storage==0.3.0
boto3==1.34.34
botocore==1.34.162
s3transfer==0.10.4
Pillow==10.2.0
rembg==2.0.65
onnxruntime==1.21.1
numpy==2.2.5
opencv-python-headless==4.11.0.86
scikit-image==0.25.2
scipy==1.15.2
PyMatting==1.1.13
tifffile==2025.3.30
imageio==2.37.0

# Caching & Performance
django-redis==5.4.0
django-cacheops==7.2
django-silk==5.1.0

# Development & Debugging
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-simple-history==3.4.0
django-import-export==3.3.6
django-cleanup==8.1.0
behave==1.2.6
coloredlogs==15.0.1
humanfriendly==10.0
tqdm==4.67.1

# Utilities
python-dotenv==1.0.0
django-environ==0.11.2
python-decouple==3.8
filetype==1.2.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.22
phonenumberslite==8.13.52
countries==0.2.0
django-ratelimit==4.1.0
beautifulsoup4==4.13.4
requests==2.32.3
stripe==7.10.0
tinify==1.7.0
qrcode==8.2
cloudinary==1.44.0
babel==2.17.0
attrs==25.3.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
referencing==0.36.2
rpds-py==0.24.0
platformdirs==4.3.7
pooch==1.8.2
protobuf==6.30.2
flatbuffers==25.2.10
networkx==3.4.2
numba==0.61.2
llvmlite==0.44.0
mpmath==1.3.0
sympy==1.14.0
lazy_loader==0.4

# Production
gunicorn==21.2.0
waitress==3.0.2
whitenoise==6.9.0

# Dependencies
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
future==1.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
jmespath==1.0.1
openapi-codec==1.3.2
packaging==24.2
pluggy==1.5.0
pyngrok==7.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
